Split restful-ws tests to sub-modules
Managing test for multiple jax-rs implemetations
using exclusions is not practical.
