+++ http/restful-ws-integration-tests/restful-ws-common/src/main/java/io/cloudevents/http/restful/ws/TestResource.java
