+++ http/restful-ws-integration-tests/restful-ws-spring/src/test/java/io/cloudevents/http/restful/ws/spring/JerseyConfig.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2018-Present The CloudEvents Authors
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package io.cloudevents.http.restful.ws.spring;
+
+import io.cloudevents.http.restful.ws.CloudEventsProvider;
+import io.cloudevents.http.restful.ws.TestResource;
+import org.glassfish.jersey.server.ResourceConfig;
+import org.springframework.context.annotation.Configuration;
+
+import javax.ws.rs.ApplicationPath;
+
+@Configuration
+public class JerseyConfig extends ResourceConfig {
+    public JerseyConfig() {
+        registerClasses(CloudEventsProvider.class, TestResource.class);
+    }
+}
