+++ pom.xml
@@ -67,7 +67,7 @@
         <module>formats/json-jackson</module>
         <module>http/vertx</module>
         <module>http/restful-ws</module>
-        <module>http/restful-ws-spring</module>
+        <module>http/restful-ws-integration-tests</module>
         <module>kafka</module>
     </modules>
